DEV: buddyMM

+++ Kernel/include/memoryManager.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 #include <stddef.h>
 
-void minit(void * start, size_t size);
-void * malloc(size_t size);
+void minit(void * start, uint64_t size);
+void * malloc(uint64_t size);
 void free(void *ptr);
-void * calloc(size_t nmemb, size_t size);
-void * realloc(void *ptr, size_t size);
+void * calloc(uint64_t nmemb, uint64_t size);
+void * realloc(void *ptr, uint64_t size);
 
 #endif//__MEMORY_MANAGER_H
